+++ packages/waggle/src/libp2p/connectionsManager.test.ts
@@ -1,6 +1,7 @@
 import { ConnectionsManager } from './connectionsManager'
 import { DummyIOServer, getPorts } from '../utils'
 import { createTmpDir, TmpDir, tmpZbayDirPath } from '../testUtils'
+import PeerId from 'peer-id'
 
 let tmpDir: TmpDir
 let tmpAppDataPath: string
@@ -53,5 +54,28 @@ describe('Connections manager', () => {
     const torControl = connectionsManager.tor.torControl
     expect(torControl.password).toEqual(torPassword)
     expect(torControl.params.port).toEqual(ports.controlPort)
+  })
+
+  it('create network', async() => {
+    const ports = await getPorts()
+    connectionsManager = new ConnectionsManager({
+      agentHost: 'localhost',
+      agentPort: ports.socksPort,
+      io: new DummyIOServer(),
+      options: {
+        env: {
+          appDataPath: tmpAppDataPath
+        },
+        torControlPort: ports.controlPort
+      }
+    })
+    await connectionsManager.init()
+    const network = await connectionsManager.createNetwork()
+    expect(network.hiddenService.onionAddress).toHaveLength(56)
+    expect(network.hiddenService.privateKey).toHaveLength(99)
+    const peerId = await PeerId.createFromJSON(network.peerId)
+    console.log(peerId, 'NETWOK')
+    expect(PeerId.isPeerId(peerId)).toBeTruthy()
+    await connectionsManager.tor.kill()
   })
 })