+++ packages/waggle/src/socket/IOProxy.ts
@@ -132,6 +132,11 @@ export default class IOProxy {
     this.io.emit(EventTypesResponse.CERTIFICATE_REGISTRATION_ERROR, { payload: message })
   }
 
+  public async createNetwork(communityId: string) {
+    const network = await this.connectionsManager.createNetwork()
+    this.io.emit(EventTypesResponse.NETWORK, { id: communityId, payload: network })
+  }
+
   public async createCommunity(communityId: string, certs: CertsData, rootCert?: string, rootKey?: string) {
     const communityData = await this.communities.create(certs)
     if (rootCert && rootKey) {