+++ packages/waggle/src/socket/constantsReponse.ts
@@ -15,5 +15,6 @@ export enum EventTypesResponse {
   NEW_COMMUNITY = 'newCommunity',
   COMMUNITY = 'community',
   REGISTRAR = 'registrar',
-  REGISTRAR_ERROR = 'registrarError'
+  REGISTRAR_ERROR = 'registrarError',
+  NETWORK = 'network'
 }